added facet count
