+++ src/test/java/com/safeway/j4u/emju/offers/api/cucumberStepDefinitions/FacetCountGrandChildandGrnadChildPlusScenarios.java
@@ -1,0 +1,66 @@
+package com.safeway.j4u.emju.offers.api.cucumberStepDefinitions;
+
+import com.safeway.j4u.emju.offers.api.framework.support.common.BaseApiTest;
+import com.safeway.j4u.emju.offers.api.framework.support.constants.GlobalConstants;
+import com.safeway.j4u.emju.offers.api.framework.support.constants.ResourceEndpointUri;
+import cucumber.api.java.en.When;
+import org.json.simple.JSONObject;
+
+import java.io.UnsupportedEncodingException;
+import java.net.URLEncoder;
+import java.util.List;
+
+public class FacetCountGrandChildandGrnadChildPlusScenarios extends BaseApiTest {
+
+
+    @When("^Business User searches offers using Startdate & OfferProgramcode & offerStatus$")
+    public void testGetOfferSearchByStartdateAndOfferProgramCodeAndOfferStatus(List<String> methodName) throws UnsupportedEncodingException {
+        testCaseName = methodName.get(0);
+        currentTsDataRowsJsonArray =
+                dataExtractor.getJSONParseTestData(
+                        GlobalConstants.testEnvironment,
+                        GlobalConstants.currentTestAPIGroupName,
+                        GlobalConstants.currentTestAPI,
+                        testCaseName);
+        currentTsJsonObject = (JSONObject) currentTsDataRowsJsonArray.get(0);
+        queryParams.put("q", currentTsJsonObject.get("q"));
+        headerParams.put(GlobalConstants.CACHECONTROL, "no-cache");
+        authenticatedGalleryEndpoint +=
+                ResourceEndpointUri.QUERY_DELIMITER +URLEncoder.encode(currentTsJsonObject.get("q").toString(),"UTF-8");
+    }
+
+    @When("^Business User searches offers using PostalCode and StoreId and Enddate$")
+    public void testGetOfferSearchByPostalCodeAndStoreIdAndEnddate(List<String> methodName) throws UnsupportedEncodingException {
+        testCaseName = methodName.get(0);
+        currentTsDataRowsJsonArray =
+                dataExtractor.getJSONParseTestData(
+                        GlobalConstants.testEnvironment,
+                        GlobalConstants.currentTestAPIGroupName,
+                        GlobalConstants.currentTestAPI,
+                        testCaseName);
+        currentTsJsonObject = (JSONObject) currentTsDataRowsJsonArray.get(0);
+        queryParams.put("q", currentTsJsonObject.get("q"));
+        headerParams.put(GlobalConstants.CACHECONTROL, "no-cache");
+        authenticatedGalleryEndpoint +=
+               ResourceEndpointUri.QUERY_DELIMITER +URLEncoder.encode(currentTsJsonObject.get("q").toString(),"UTF-8");
+    }
+
+    @When("^Business User searches offers using StartDate and EndDate and PostalCd and OfferStatus and StoreId and OfferProgramcode$")
+    public void testGetOfferSearchByStartDateAndEndDateAndPostalCodeAndOfferStatusAndStoreIdAndOfferProgramcode(List<String> methodName) throws UnsupportedEncodingException {
+        testCaseName = methodName.get(0);
+        currentTsDataRowsJsonArray =
+                dataExtractor.getJSONParseTestData(
+                        GlobalConstants.testEnvironment,
+                        GlobalConstants.currentTestAPIGroupName,
+                        GlobalConstants.currentTestAPI,
+                        testCaseName);
+        currentTsJsonObject = (JSONObject) currentTsDataRowsJsonArray.get(0);
+        queryParams.put("q", currentTsJsonObject.get("q"));
+        headerParams.put(GlobalConstants.CACHECONTROL, "no-cache");
+        authenticatedGalleryEndpoint +=
+                ResourceEndpointUri.QUERY_DELIMITER +URLEncoder.encode(currentTsJsonObject.get("q").toString(),"UTF-8");
+    }
+
+
+
+}
